Store faithful json data in allocator_channel_commitments.app_attributes column
For the server wallet to manage state for a non-perfect-information game
(such as a reveal-commit implementation of rock paper scissors) it's
necessary for the wallet to store application data that would not be
sent across the wire to the opponent, for instance.

In order to keep the wallet application-agnostic, this change adds
a `sanitize` function, which the application should pass to the wallet
when the wallet is initialized.

Whenever an internal commitment is cast to a core Commitment, the
sanitize method is called on the application attributes. This is
consistent with the convention that a `Commitment` is exactly the
information that should be passed between parties.

An incidental change is that the application attributes are stored as
a json column, rather than encoded as bytes. This precludes the
requirement for the app to provide a sanitize method casting from
structured attributes to unfaithful bytes, and a separate, faithful
method of casting to bytes. It has the side effect of being
developer-friendly -- debugging and testing is much nicer with
structured data.